+++ service/history/api/getworkflowexecutionhistory/api.go
@@ -37,7 +37,6 @@ import (
 	persistencespb "go.temporal.io/server/api/persistence/v1"
 	tokenspb "go.temporal.io/server/api/token/v1"
 	"go.temporal.io/server/common"
-	"go.temporal.io/server/common/failure"
 	"go.temporal.io/server/common/headers"
 	"go.temporal.io/server/common/namespace"
 	"go.temporal.io/server/common/persistence/transitionhistory"
@@ -198,9 +197,10 @@ func Invoke(
 	var historyBlob []*commonpb.DataBlob
 	config := shardContext.GetConfig()
 	sendRawHistoryBetweenInternalServices := config.SendRawHistoryBetweenInternalServices()
+	sendRawWorkflowHistoryForNamespace := config.SendRawWorkflowHistory(request.Request.GetNamespace())
 	if isCloseEventOnly {
 		if !isWorkflowRunning {
-			if config.SendRawWorkflowHistory(request.Request.GetNamespace()) || sendRawHistoryBetweenInternalServices {
+			if sendRawWorkflowHistoryForNamespace || sendRawHistoryBetweenInternalServices {
 				historyBlob, _, err = api.GetRawHistory(
 					ctx,
 					shardContext,
@@ -254,7 +254,7 @@ func Invoke(
 				continuationToken = nil
 			}
 		} else {
-			if config.SendRawWorkflowHistory(request.Request.GetNamespace()) || sendRawHistoryBetweenInternalServices {
+			if sendRawWorkflowHistoryForNamespace || sendRawHistoryBetweenInternalServices {
 				historyBlob, continuationToken.PersistenceToken, err = api.GetRawHistory(
 					ctx,
 					shardContext,
@@ -303,7 +303,7 @@ func Invoke(
 
 	// if SendRawHistoryBetweenInternalServices is enabled, we do this check in frontend service
 	if len(history.Events) > 0 {
-		err = FixFollowEvents(ctx, versionChecker, isCloseEventOnly, history)
+		err = api.FixFollowEvents(ctx, versionChecker, isCloseEventOnly, history)
 		if err != nil {
 			return nil, err
 		}
@@ -314,19 +314,10 @@ func Invoke(
 	// send history in raw format in History field of historyservice.GetWorkflowExecutionHistoryResponseWithRaw.
 	// If SendRawWorkflowHistory is enabled for this namespace, raw history will be appended to RawHistory field in
 	// workflowservice.GetWorkflowExecutionHistoryResponse.
-	if sendRawHistoryBetweenInternalServices && !config.SendRawWorkflowHistory(request.Request.GetNamespace()) {
+	if sendRawHistoryBetweenInternalServices && !sendRawWorkflowHistoryForNamespace {
 		rawHistory = make([][]byte, 0, len(historyBlob))
 		for _, blob := range historyBlob {
 			rawHistory = append(rawHistory, blob.Data)
-		}
-		// If there are no events in the history, frontend will not be able to deserialize the response to History object.
-		// In that case, create an empty history object and set it in the response.
-		if len(rawHistory) == 0 {
-			blob, err := history.Marshal()
-			if err != nil {
-				return nil, err
-			}
-			rawHistory = [][]byte{blob}
 		}
 		historyBlob = nil
 	}
@@ -339,91 +330,5 @@ func Invoke(
 			Archived:      false,
 		},
 		History: rawHistory,
-	}, nil
-}
-
-func FixFollowEvents(
-	ctx context.Context,
-	versionChecker headers.VersionChecker,
-	isCloseEventOnly bool,
-	history *historypb.History,
-) error {
-	// Backwards-compatibility fix for retry events after #1866: older SDKs don't know how to "follow"
-	// subsequent runs linked in WorkflowExecutionFailed or TimedOut events, so they'll get the wrong result
-	// when trying to "get" the result of a workflow run. (This applies to cron runs also but "get" on a cron
-	// workflow isn't really sensible.)
-	//
-	// To handle this in a backwards-compatible way, we'll pretend the completion event is actually
-	// ContinuedAsNew, if it's Failed or TimedOut. We want to do this only when the client is looking for a
-	// completion event, and not when it's getting the history to display for other purposes. The best signal
-	// for that purpose is `isCloseEventOnly`. (We can't use `isLongPoll` also because in some cases, older
-	// versions of the Java SDK don't set that flag.)
-	//
-	// TODO: We can remove this once we no longer support SDK versions prior to around September 2021.
-	// Revisit this once we have an SDK deprecation policy.
-	followsNextRunId := versionChecker.ClientSupportsFeature(ctx, headers.FeatureFollowsNextRunID)
-	if isCloseEventOnly && !followsNextRunId && len(history.Events) > 0 {
-		lastEvent := history.Events[len(history.Events)-1]
-		fakeEvent, err := makeFakeContinuedAsNewEvent(ctx, lastEvent)
-		if err != nil {
-			return err
-		}
-		if fakeEvent != nil {
-			history.Events[len(history.Events)-1] = fakeEvent
-		}
-	}
-	return nil
-}
-
-func makeFakeContinuedAsNewEvent(
-	_ context.Context,
-	lastEvent *historypb.HistoryEvent,
-) (*historypb.HistoryEvent, error) {
-	switch lastEvent.EventType {
-	case enumspb.EVENT_TYPE_WORKFLOW_EXECUTION_COMPLETED:
-		if lastEvent.GetWorkflowExecutionCompletedEventAttributes().GetNewExecutionRunId() == "" {
-			return nil, nil
-		}
-	case enumspb.EVENT_TYPE_WORKFLOW_EXECUTION_FAILED:
-		if lastEvent.GetWorkflowExecutionFailedEventAttributes().GetNewExecutionRunId() == "" {
-			return nil, nil
-		}
-	case enumspb.EVENT_TYPE_WORKFLOW_EXECUTION_TIMED_OUT:
-		if lastEvent.GetWorkflowExecutionTimedOutEventAttributes().GetNewExecutionRunId() == "" {
-			return nil, nil
-		}
-	default:
-		return nil, nil
-	}
-
-	// We need to replace the last event with a continued-as-new event that has at least the
-	// NewExecutionRunId field. We don't actually need any other fields, since that's the only one
-	// the client looks at in this case, but copy the last result or failure from the real completed
-	// event just so it's clear what the result was.
-	newAttrs := &historypb.WorkflowExecutionContinuedAsNewEventAttributes{}
-	switch lastEvent.EventType {
-	case enumspb.EVENT_TYPE_WORKFLOW_EXECUTION_COMPLETED:
-		attrs := lastEvent.GetWorkflowExecutionCompletedEventAttributes()
-		newAttrs.NewExecutionRunId = attrs.NewExecutionRunId
-		newAttrs.LastCompletionResult = attrs.Result
-	case enumspb.EVENT_TYPE_WORKFLOW_EXECUTION_FAILED:
-		attrs := lastEvent.GetWorkflowExecutionFailedEventAttributes()
-		newAttrs.NewExecutionRunId = attrs.NewExecutionRunId
-		newAttrs.Failure = attrs.Failure
-	case enumspb.EVENT_TYPE_WORKFLOW_EXECUTION_TIMED_OUT:
-		attrs := lastEvent.GetWorkflowExecutionTimedOutEventAttributes()
-		newAttrs.NewExecutionRunId = attrs.NewExecutionRunId
-		newAttrs.Failure = failure.NewTimeoutFailure("workflow timeout", enumspb.TIMEOUT_TYPE_START_TO_CLOSE)
-	}
-
-	return &historypb.HistoryEvent{
-		EventId:   lastEvent.EventId,
-		EventTime: lastEvent.EventTime,
-		EventType: enumspb.EVENT_TYPE_WORKFLOW_EXECUTION_CONTINUED_AS_NEW,
-		Version:   lastEvent.Version,
-		TaskId:    lastEvent.TaskId,
-		Attributes: &historypb.HistoryEvent_WorkflowExecutionContinuedAsNewEventAttributes{
-			WorkflowExecutionContinuedAsNewEventAttributes: newAttrs,
-		},
 	}, nil
 }
